moved collaborators to production files
