+++ TriviaEngine/TriviaEngine/QuestionsLoader.swift
@@ -1,0 +1,3 @@
+public protocol QuestionsLoader {
+    func load() throws -> [Question]
+}
